requirements.txt
Django
gunicorn
psycopg2-binary
assemblyai
groq
nltk
scikit-learn
numpy
scipy
python-dotenv
yt-dlp